requirements.txt
amqp==2.6.1
asgiref==3.3.1
billiard==3.6.3.0
celery==4.4.7
coverage==5.3
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.4
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-cleanup==5.1.0
django-cors-headers==3.5.0
django-dbbackup==3.3.0
django-debug-toolbar==3.2
django-filter==2.4.0
django-import-export==2.4.0
django-jazzmin==2.4.1
django-prometheus==2.1.0
django-timezone-field==4.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
et-xmlfile==1.0.1
flake8==3.8.4
future==0.18.2
jdcal==1.4.1
kombu==4.6.11
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.0.1
pipdeptree==1.0.0
prometheus-client==0.9.0
psycopg2==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
